% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/birthday.R
\name{birthday}
\alias{birthday}
\title{Birthday function}
\usage{
birthday(x)
}
\arguments{
\item{x}{class size}
}
\value{
probability of people with the same bdays from a class w/ x amount of students
}
\description{
Birthday function
}
\examples{
birthday(1:10)
}
