% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Fish in Tennessee River}
\format{
## `ddt`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{RIVER}{River type}
  \item{MILE}{number}
  \item{SPECIES}{fish name}
  \item{LENGTH}{lrngth in cm}
  \item{WEIGHT}{number}
  \item{DDT}{number}
  ...
}
}
\source{
MS
}
\usage{
ddt
}
\description{
A subset of data from the World Health Organization Global Tuberculosis
Report ...
}
\keyword{datasets}
