% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{FIREDAM Data}
\format{
## `fire`
data of a 15 x 2:
\describe{
  \item{DISTANCE}{distance from fire station in miles}
  \item{DAMAGE}{x 1000 dolllars in damage}

  ...
}
}
\source{
<>
}
\usage{
fire
}
\description{
FIREDAM Data
}
\keyword{datasets}
