% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{My Binomial that shows a graph and table}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{iterations or trials}

\item{n}{size or range of numbers ex. n=5 5 numbers show up on x axis of plot}

\item{p}{probability}
}
\value{
a barplot showing the amount of times a number shows up over a
certain amount of iterations also shows a table for the amount of success
}
\description{
My Binomial that shows a graph and table
}
\examples{
mybin(iter=100, n=10, p=0.5)
}
