% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Bootstrap Function}
\usage{
myboot2(iter = 10000, x, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations}

\item{x}{a sample of an r version of a distribution function (sam=rnorm(25,mean=25,sd=10))}

\item{fun}{the function to use (mean, var, median, sd, quantile, IQR, or a configured one)}

\item{alpha}{alpha (remainder of p, exa: p=0.95, alpha=0.05)}

\item{cx}{character expansion}

\item{...}{aditional parameters (does not mean excetera)}
}
\value{
a bootstrap distribution with a confidence interval in plot form
}
\description{
Bootstrap Function
}
\examples{
sam=rnorm(25,mean=25,sd=10)
myboot2(iter=10000, x=sam, fun='var', alpha=0.2)
}
