% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myclt.R
\name{myclt}
\alias{myclt}
\title{My Central Limit Theorem (CLT) Uniform Function}
\usage{
myclt(n, iter)
}
\arguments{
\item{n}{size of sample}

\item{iter}{iterations}
}
\value{
a histogram of the CLT uniform with every w value
}
\description{
My Central Limit Theorem (CLT) Uniform Function
}
\examples{
myclt(n=10, iter=10000)

}
