% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlik.R
\name{mymaxlik}
\alias{mymaxlik}
\title{Maximum Likelihood}
\usage{
mymaxlik(lfun, x, param, ...)
}
\arguments{
\item{lfun}{a log function created separately}

\item{x}{list of variables}

\item{param}{parameters for the plot}

\item{...}{does not mean excetera}
}
\value{
a plot of the maximum likelihood for a specific distribution.
This also prints out the first maximum value, the parameter at i, yi, and the slope.
}
\description{
Maximum Likelihood
}
\examples{
logbin=function(x,param) log(dbinom(x,prob=param,size=20))
mymaxlik(x=c(3,3,4,3,4,5,5,4),param=seq(0,1,length=1000),
         lfun=logbin,xlab=expression(pi),main="Binomial",cex.main=2)
}
