% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Normal Curve Function}
\usage{
myncurve(a, mu, sigma)
}
\arguments{
\item{a}{a value of P(X<=a)}

\item{mu}{mean}

\item{sigma}{sd}
}
\value{
plot of normal distribution w/ shaded area of P(X<=a) and displays the probability in the command line
}
\description{
Normal Curve Function
}
\examples{
myncurve(10,10,4)
}
