% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ntickets.R
\name{ntickets}
\alias{ntickets}
\title{Amount of tickets bought on the plane (n)}
\usage{
ntickets(N, g, p)
}
\arguments{
\item{N}{number of seats total}

\item{g}{gamma or pain (more people than seats)}

\item{p}{probability people will show up}
}
\value{
A list of variables (total tickets bought using discrete (nd)
and continuous (nc) methods, N, g, and p) and tables of Objective vs n
for discrete and continuous
}
\description{
Amount of tickets bought on the plane (n)
}
\examples{
ntickets(N=200, g=0.02, p=0.95)
}
