% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Fish in Tennessee River}
\format{
## `ddt`
A data frame with 7,240 rows and 60 columns:
\describe{
  \item{RIVER}{River found}
  \item{MILE}{distance in miles}
  \item{SPECIES}{fish name}
  \item{LENGTH}{length of fish in cm}
  \item{WEIGHT}{weight of fish in grams}
  \item{DDT}{parts per million}
  ...
}
}
\source{
Statistics for Engineering and the Sciences, Sixth Edition (MS)
}
\usage{
ddt
}
\description{
Characteristics of Contaminated Fish in the Tennessee River
}
\keyword{datasets}
