% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire Damage Data}
\format{
## `fire`
data of a 15 x 2:
\describe{
  \item{DISTANCE}{distance from fire station in miles}
  \item{DAMAGE}{x1000 dollars in damage}

  ...
}
}
\source{
Statistics for Engineering and the Sciences, Sixth Edition (MS)
}
\usage{
fire
}
\description{
Data of fire damage (in dollars) in major residential fires related to the distance between the residence and fire station
}
\keyword{datasets}
